Relocate BaseMVAValueMapProducer to PhysicsTools/PatAlgos
